Moved shrinkage functions to utils
